refactor(derive): split enum_derive into per-backend modules
Replace the monolithic enum_derive.rs (658 lines) with a module layout:

- mod.rs: shared parsing, macros, and input struct
- backend.rs: common token emission (gated on postgres|mysql)
- postgres.rs: PostgreSQL-specific code generation
- mysql.rs: MySQL-specific code generation

Also add dead_code suppression annotations on the shared path so
compilation is clean when no backend feature is enabled.
